migrate packageParser, log updates, fix package override
