+++ packages/rnv/src/projectTools/buildHooks.js
@@ -34,7 +34,7 @@ const rnvHooksRun = c => new Promise((resolve, reject) => {
 });
 
 const executePipe = async (c, key) => {
-    logTask(`executePipe:${key}`);
+    logTask('executePipe', `${key}`);
 
     const pipesConfig = c.buildConfig?.pipes;
     if (!pipesConfig || (pipesConfig && pipesConfig.includes(key))) {
